requirements.txt
arxiv==2.1.0
beautifulsoup4==4.12.3
boto3==1.26.76
fastapi==0.110.0
langchain==0.1.13
langchain_openai==0.1.1
langchain_pinecone==0.0.3
openai==1.14.3
pinecone-client==3.0.0
psycopg2-binary
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
redis==4.5.5
requests==2.31.0
tiktoken==0.5.2
tqdm
uvicorn==0.21.1
ffmpeg-python==0.2.0
google-generativeai==0.5.2
rake-nltk
cohere
paperswithcode-client
celery==5.4.0
PyMuPDF==1.24.9